Move old managers aside
