+++ scheduler/managers.py
@@ -6,8 +6,7 @@ from django.db import models
 
 from django.utils import timezone
 
-from organizations.models import Facility
-from places.models import Place, Area, Region, Country
+from scheduler.old_managers import NeedManager
 
 
 class EnrolmentManager(models.Manager):
@@ -27,49 +26,6 @@ class EnrolmentManager(models.Manager):
         query_set = query_set.exclude(shift__start_time__gte=graced_end,
                                       shift__end_time__gte=graced_end)
         return query_set
-
-
-# This "second" view allows to create/publish the shifts for the X coming days out of the
-# recurring events.
-# On REST side, it should be possible to request all events from the template for a specified day,
-# return them to the front-end which will publish them and/or modify them via PUTs statements
-
-
-class NeedManager(models.Manager):
-    def at_facility(self, facility):
-        return self.get_queryset().filter(facility=facility)
-
-    def at_place(self, place):
-        return self.get_queryset().filter(facility__place=place)
-
-    def in_area(self, area):
-        return self.get_queryset().filter(facility__place__area=area)
-
-    def in_region(self, region):
-        return self.get_queryset().filter(facility__place__area__region=region)
-
-    def in_country(self, country):
-        return self.get_queryset().filter(
-            facility__place__area__region__country=country)
-
-    def by_geography(self, geo_affiliation):
-        if isinstance(geo_affiliation, Facility):
-            return self.at_facility(geo_affiliation)
-        elif isinstance(geo_affiliation, Place):
-            return self.at_place(geo_affiliation)
-        elif isinstance(geo_affiliation, Area):
-            return self.in_area(geo_affiliation)
-        elif isinstance(geo_affiliation, Region):
-            return self.in_region(geo_affiliation)
-        elif isinstance(geo_affiliation, Country):
-            return self.in_country(geo_affiliation)
-
-
-class OpenNeedManager(NeedManager):
-    def get_queryset(self):
-        now = timezone.now()
-        qs = super(OpenNeedManager, self).get_queryset()
-        return qs.filter(ending_time__gte=now)
 
 
 class ShiftManager(NeedManager):
@@ -116,20 +72,4 @@ class RecurringEventManager(models.Manager):
                            disabled=False)
 
 
-class ShiftHelperManager(models.Manager):
-    def conflicting(self, need, user_account=None, grace=timedelta(hours=1)):
 
-        grace = grace or timedelta(0)
-        graced_start = need.starting_time + grace
-        graced_end = need.ending_time - grace
-
-        query_set = self.get_queryset().select_related('need', 'user_account')
-
-        if user_account:
-            query_set = query_set.filter(user_account=user_account)
-
-        query_set = query_set.exclude(need__starting_time__lt=graced_start,
-                                      need__ending_time__lte=graced_start)
-        query_set = query_set.exclude(need__starting_time__gte=graced_end,
-                                      need__ending_time__gte=graced_end)
-        return query_set
